Add VM Priority input to Terraform workflow
This commit adds a new input, "vm-priority", to the Terraform workflow. The input allows users to specify the priority of the virtual machine, with options for "Spot" or "Regular". The value of the input is then passed to the "TF_VAR_vm_priority" variable in the workflow. Additionally, the "priority" and "eviction_policy" attributes in the "azurerm_linux_virtual_machine" resource in the "main.tf" file are updated to use the "vm_priority" variable.